Add arbitrary template extensions and rendered filenames.
This commit adds the ability to use extensions on templates based on the
global configuration variable `template_extensions`. Complementing this
change is the ability to render content with a different filename than
`index.html` with the frontmatter variable `filename`, which by default
will still render the page as `index.html` at its given path.